+++ index.html
@@ -166,5 +166,5 @@
   </template>
 
 </body>
-<script src="./js/main.js"></script>
+<script type="module" src="./js/main.js"></script>
 </html>
